// lib/types.generated.ts
type Simplify<T> = { [KeyType in keyof T]: T[KeyType] };
// ...
export interface BlogPostDocumentDataGuestsItem {
	/**
	 * Name field in *Blog Post → Guests*
	 *
	 * - **Field Type**: Text
	 * - **Placeholder**: *None*
	 * - **API ID Path**: blog_post.guests[].name
	 * - **Documentation**: https://prismic.io/docs/field#key-text
	 */
	name: prismic.KeyTextField;
}
// ...
type BlogPostDocumentDataBodySlice = BlogPostDocumentDataBodyTextSlice | BlogPostDocumentDataBodyImageSlice | BlogPostDocumentDataBodyIframeSlice | BlogPostDocumentDataBodyHtmlSlice | BlogPostDocumentDataBodyHighlightedTextSlice

/**
 * Content for Blog Post documents
 */
interface BlogPostDocumentData {
	/**
	 * Headline field in *Blog Post*
	 *
	 * - **Field Type**: Title
	 * - **Placeholder**: *None*
	 * - **API ID Path**: blog_post.headline
	 * - **Tab**: Main
	 * - **Documentation**: https://prismic.io/docs/field#rich-text-title
	 */
	headline: prismic.TitleField;
	
	/**
	 * Excerpt field in *Blog Post*
	 *
	 * - **Field Type**: Rich Text
	 * - **Placeholder**: *None*
	 * - **API ID Path**: blog_post.excerpt
	 * - **Tab**: Main
	 * - **Documentation**: https://prismic.io/docs/field#rich-text-title
	 */
	excerpt: prismic.RichTextField;
	
	/**
	 * Type field in *Blog Post*
	 *
	 * - **Field Type**: Select
	 * - **Placeholder**: *None*
	 * - **Default Value**: Blog
	 * - **API ID Path**: blog_post.type
	 * - **Tab**: Main
	 * - **Documentation**: https://prismic.io/docs/field#select
	 */
	type: prismic.SelectField<"Blog" | "Podcast", "filled">;
	
	/**
	 * Image field in *Blog Post*
	 *
	 * - **Field Type**: Image
	 * - **Placeholder**: *None*
	 * - **API ID Path**: blog_post.cover_image
	 * - **Tab**: Main
	 * - **Documentation**: https://prismic.io/docs/field#image
	 */
	cover_image: prismic.ImageField<"main2x" | "preview" | "preview2x">;
	
	/**
	 * Image Text field in *Blog Post*
	 *
	 * - **Field Type**: Rich Text
	 * - **Placeholder**: *None*
	 * - **API ID Path**: blog_post.cover_image_text
	 * - **Tab**: Main
	 * - **Documentation**: https://prismic.io/docs/field#rich-text-title
	 */
	cover_image_text: prismic.RichTextField;
	
	/**
	 * Show text bellow image field in *Blog Post*
	 *
	 * - **Field Type**: Boolean
	 * - **Placeholder**: *None*
	 * - **Default Value**: false
	 * - **API ID Path**: blog_post.show_text_bellow_image
	 * - **Tab**: Main
	 * - **Documentation**: https://prismic.io/docs/field#boolean
	 */
	show_text_bellow_image: prismic.BooleanField;
	
	/**
	 * Embed field in *Blog Post*
	 *
	 * - **Field Type**: Embed
	 * - **Placeholder**: Embed video or other oembed compatible urls
	 * - **API ID Path**: blog_post.embed
	 * - **Tab**: Main
	 * - **Documentation**: https://prismic.io/docs/field#embed
	 */
	embed: prismic.EmbedField
	
	/**
	 * Author field in *Blog Post*
	 *
	 * - **Field Type**: Content Relationship
	 * - **Placeholder**: *None*
	 * - **API ID Path**: blog_post.author
	 * - **Tab**: Main
	 * - **Documentation**: https://prismic.io/docs/field#link-content-relationship
	 */
	author: prismic.ContentRelationshipField<"authors">;
	
	/**
	 * Publishing date field in *Blog Post*
	 *
	 * - **Field Type**: Date
	 * - **Placeholder**: *None*
	 * - **API ID Path**: blog_post.publishing_date
	 * - **Tab**: Main
	 * - **Documentation**: https://prismic.io/docs/field#date
	 */
	publishing_date: prismic.DateField;
	
	/**
	 * Guests field in *Blog Post*
	 *
	 * - **Field Type**: Group
	 * - **Placeholder**: *None*
	 * - **API ID Path**: blog_post.guests[]
	 * - **Tab**: Main
	 * - **Documentation**: https://prismic.io/docs/field#group
	 */
	guests: prismic.GroupField<Simplify<BlogPostDocumentDataGuestsItem>>;
	
	/**
	 * Slice zone field in *Blog Post*
	 *
	 * - **Field Type**: Slice Zone
	 * - **Placeholder**: *None*
	 * - **API ID Path**: blog_post.body[]
	 * - **Tab**: Main
	 * - **Documentation**: https://prismic.io/docs/field#slices
	 */
	body: prismic.SliceZone<BlogPostDocumentDataBodySlice>;/**
	 * Title field in *Blog Post*
	 *
	 * - **Field Type**: Text
	 * - **Placeholder**: *None*
	 * - **API ID Path**: blog_post.seo_title
	 * - **Tab**: SEO
	 * - **Documentation**: https://prismic.io/docs/field#key-text
	 */
	seo_title: prismic.KeyTextField;
	
	/**
	 * Description field in *Blog Post*
	 *
	 * - **Field Type**: Text
	 * - **Placeholder**: *None*
	 * - **API ID Path**: blog_post.seo_description
	 * - **Tab**: SEO
	 * - **Documentation**: https://prismic.io/docs/field#key-text
	 */
	seo_description: prismic.KeyTextField;
	
	/**
	 * Social Media Image field in *Blog Post*
	 *
	 * - **Field Type**: Image
	 * - **Placeholder**: *None*
	 * - **API ID Path**: blog_post.social_media_image
	 * - **Tab**: SEO
	 * - **Documentation**: https://prismic.io/docs/field#image
	 */
	social_media_image: prismic.ImageField<never>;
}

/**
 * Blog Post document from Prismic
 *
 * - **API ID**: `blog_post`
 * - **Repeatable**: `true`
 * - **Documentation**: https://prismic.io/docs/custom-types
 *
 * @typeParam Lang - Language API ID of the document.
 */
export type BlogPostDocument<Lang extends string = string> = prismic.PrismicDocumentWithUID<Simplify<BlogPostDocumentData>, "blog_post", Lang>;